+++ crates/bevy_picking_core/src/focus.rs
@@ -136,74 +136,3 @@ fn build_hover_map(
         }
     }
 }
-
-pub fn pointer_events(
-    // Input
-    pointers: Query<(&PointerId, &PointerInteraction)>, // <- what happened last frame
-    mut input_presses: EventReader<input::InputPress>,
-    mut pointer_move_in: EventReader<input::InputMove>,
-    hover_map: Res<HoverMap>,
-    // Output
-    mut pointer_move: EventWriter<PointerMove>,
-    mut pointer_over: EventWriter<PointerOver>,
-    mut pointer_out: EventWriter<PointerOut>,
-    mut pointer_up: EventWriter<PointerUp>,
-    mut pointer_down: EventWriter<PointerDown>,
-) {
-    let input_presses: Vec<&InputPress> = input_presses.iter().collect();
-
-    for event in pointer_move_in.iter() {
-        for hover_entity in hover_map.get(&event.id()).iter().flat_map(|h| h.iter()) {
-            pointer_move.send(PointerMove::new(&event.id(), hover_entity, Move))
-        }
-    }
-
-    for (pointer_id, pointer_interaction) in pointers.iter() {
-        let just_pressed = input_presses
-            .iter()
-            .filter_map(|click| (&click.id == pointer_id).then_some(click.press))
-            .last();
-
-        // If the entity is hovered...
-        for hover_entity in hover_map.get(pointer_id).iter().flat_map(|h| h.iter()) {
-            // ...but was not hovered last frame...
-            if matches!(
-                pointer_interaction.get(hover_entity),
-                Some(Interaction::None) | None
-            ) {
-                pointer_over.send(PointerOver::new(pointer_id, hover_entity, Over));
-            }
-
-            match just_pressed {
-                Some(PressStage::Down) => {
-                    pointer_down.send(PointerDown::new(pointer_id, hover_entity, Down));
-                }
-                Some(PressStage::Up) => {
-                    pointer_up.send(PointerUp::new(pointer_id, hover_entity, Up));
-                }
-                None => (),
-            }
-        }
-
-        if let Some(hover_entities) = hover_map.get(pointer_id) {
-            // If the entity was hovered last frame...
-            for entity in pointer_interaction
-                .iter()
-                .filter_map(|(entity, interaction)| {
-                    matches!(interaction, Interaction::Hovered | Interaction::Clicked)
-                        .then_some(entity)
-                })
-            {
-                // ...but is now not being hovered...
-                if !hover_entities.contains(entity) {
-                    if matches!(just_pressed, Some(PressStage::Up)) {
-                        // ...the pointer is considered just up on this entity even though it was
-                        // not hovering the entity this frame
-                        pointer_up.send(PointerUp::new(pointer_id, entity, Up));
-                    }
-                    pointer_out.send(PointerOut::new(pointer_id, entity, Out));
-                }
-            }
-        }
-    }
-}
